requirements.txt
langchain==0.1.0
openai==1.7.2
langchain-openai==0.0.2
langchain-community==0.0.12
langchainhub==0.1.14
python-dotenv
chromadb==0.4.22
streamlit==1.36.0